pyproject.toml
[project]
name = "skill-importance-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.6.3",
    "huggingface-hub>=0.29.1",
    "ipykernel>=6.29.5",
    "langchain>=0.3.19",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.18",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.3.7",
    "matplotlib>=3.10.1",
    "numpy>=1.26.4",
    "plotly>=6.0.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.2",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.49.0",
]
